refactor: 패키지 분리 제거
