+++ src/main/java/kitchenpos/domain/OrderLineItem.java
@@ -1,7 +1,4 @@
-package kitchenpos.domain.order_line_item;
-
-import kitchenpos.domain.menu.Menu;
-import kitchenpos.domain.order.Order;
+package kitchenpos.domain;
 
 import javax.persistence.*;
