+++ src/main/java/kitchenpos/domain/OrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.order;
+package kitchenpos.domain;
 
 public enum OrderStatus {
     COOKING, MEAL, COMPLETION
