+++ src/main/java/kitchenpos/domain/OrderTable.java
@@ -1,6 +1,4 @@
-package kitchenpos.domain.order_table;
-
-import kitchenpos.domain.table_group.TableGroup;
+package kitchenpos.domain;
 
 import javax.persistence.*;
