+++ src/main/java/kitchenpos/domain/Product.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.product;
+package kitchenpos.domain;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
