+++ src/main/java/kitchenpos/domain/TableGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.table_group;
+package kitchenpos.domain;
 
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
