+++ src/main/java/kitchenpos/domain/repository/MenuRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.domain.repository;
 
-import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
