+++ src/main/java/kitchenpos/domain/repository/TableGroupRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.domain.repository;
 
-import kitchenpos.domain.table_group.TableGroup;
+import kitchenpos.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
