+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -1,6 +1,6 @@
 package kitchenpos.application;
 
-import kitchenpos.domain.product.Product;
+import kitchenpos.domain.Product;
 import kitchenpos.domain.repository.ProductRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
