+++ src/test/java/support/fixture/OrderBuilder.java
@@ -1,10 +1,6 @@
 package support.fixture;
 
-import kitchenpos.domain.order.Order;
-import kitchenpos.domain.order.OrderLineItems;
-import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.order_line_item.OrderLineItem;
-import kitchenpos.domain.order_table.OrderTable;
+import kitchenpos.domain.*;
 
 import java.util.Collections;
 import java.util.List;
