+++ src/test/java/support/fixture/TableBuilder.java
@@ -1,7 +1,7 @@
 package support.fixture;
 
-import kitchenpos.domain.order_table.OrderTable;
-import kitchenpos.domain.table_group.TableGroup;
+import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.TableGroup;
 
 public class TableBuilder {
