+++ src/test/java/support/fixture/TableGroupBuilder.java
@@ -1,8 +1,8 @@
 package support.fixture;
 
-import kitchenpos.domain.order_table.OrderTable;
-import kitchenpos.domain.table_group.OrderTables;
-import kitchenpos.domain.table_group.TableGroup;
+import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.OrderTables;
+import kitchenpos.domain.TableGroup;
 
 import java.util.ArrayList;
 import java.util.List;
